src/server/requirements.txt
#Flask Framework
Flask==2.0.1
PyJWT==2.3.0

#AWS Integration
boto3==1.18.53
awscli==1.20.53

#Other Libraries
python-dotenv==0.19.1